Added the ability to recall (from a file) what type of stage is on which axis, and then use this information to set the correct PID parameters plus velocity etc, from a stage database.
Added a function that gets the limit status. If a limit switch is reached in 
the process of moving, then motion is stopped (by the hardware) and further 
motion in that direction is prevented (again by the hardware). All the software
has to do is to determine what the status of the limit switches is, and then
the user can determine which direction we need to move in.

Small changes to manual_usb_stage to use the above additions.